Merged: oleh_derevenko#1927: Cosmetic: Explicit cast of INFINITY to double added for configuration with doubles to solve compile error

+++ include/ode/odeconfig.h
@@ -95,7 +95,7 @@
   #ifdef dSINGLE
     #define dInfinity ((float)INFINITY)
   #else
-    #define dInfinity INFINITY
+    #define dInfinity ((double)INFINITY)
   #endif
 #elif defined(HUGE_VAL)
   #ifdef dSINGLE
